fix(dashboard): few cleanups
