+++ app/components/ui/index.ts
@@ -1,7 +1,6 @@
 export { default as Button } from "./Button";
 export { default as Field } from "./Field";
 export { default as Select } from "./Select";
-export type { SelectOption } from "./Select";
 export { default as MultiSelect } from "./MultiSelect";
 export { default as Modal } from "./Modal";
 export { default as Loader, LoaderBox } from "./Loader";
