+++ app/lib/types/ui.ts
@@ -1,0 +1,4 @@
+export interface SelectOption {
+  value: string;
+  label: string;
+}
